Add files via upload
This is a senior capstone project. 
How it work:
This has a loop of basic advertisements and it keeps displaying the advertisements in the loop. Once any brand from 'brand_list' is detected, it breaks out the loop and display the advertisement of the detected brand. When the detected brand advertisement is done, program goes back to the basic advertisement loop and keep looking for logo to detect. 
It is using CLIP - ViT-B-32-quickgelu and laion400m_e32
How to run:
you have to run this in python 3.10 (3.11 is the latest). 
To run, download libraries first, on CMD, 'pip install open_clip_torch' 'pip install torch' 'pip install opencv-python' 'pip install numpy' 
On VS code, open terminal change Power Shell to CMD, change the directory and 'python main.py' to run the code
when first time running the code on a device, it will download some pre-trained variables (only first time on a device)
As it is running it will save the brand name on 'brand_queue'.
press 'esc' button to exit the program. then it will show you what is in 'brand_queue' on terminal